Split ReplicatorBuilder.js into seperate files
